feat(ivy): support injecting the injector (#26699)
PR Close #26699
