+++ packages/core/src/di.ts
@@ -16,7 +16,8 @@ export * from './di/metadata';
 export {InjectableType, InjectorType, defineInjectable, defineInjector} from './di/defs';
 export {forwardRef, resolveForwardRef, ForwardRefFn} from './di/forward_ref';
 export {Injectable, InjectableDecorator, InjectableProvider} from './di/injectable';
-export {inject, InjectFlags, INJECTOR, Injector} from './di/injector';
+export {INJECTOR, Injector} from './di/injector';
+export {inject, InjectFlags} from './di/injector_compatibility';
 export {ReflectiveInjector} from './di/reflective_injector';
 export {StaticProvider, ValueProvider, ConstructorSansProvider, ExistingProvider, FactoryProvider, Provider, TypeProvider, ClassProvider} from './di/provider';
 export {createInjector} from './di/r3_injector';
