+++ packages/core/src/render3/interfaces/injector.ts
@@ -7,7 +7,7 @@
  */
 
 import {InjectionToken} from '../../di/injection_token';
-import {InjectFlags} from '../../di/injector';
+import {InjectFlags} from '../../di/injector_compatibility';
 import {Type} from '../../type';
 import {TElementNode} from './node';
 import {LViewData, TData} from './view';
