+++ packages/core/test/bundling/animation_world/bundle.golden_symbols.json
@@ -48,7 +48,7 @@
     "name": "DefaultIterableDifferFactory"
   },
   {
-    "name": "EMPTY$1"
+    "name": "EMPTY"
   },
   {
     "name": "EMPTY_ARR"
@@ -81,7 +81,7 @@
     "name": "HOST_NODE"
   },
   {
-    "name": "INJECTOR$1"
+    "name": "INJECTOR"
   },
   {
     "name": "INJECTOR_SIZE"
@@ -138,7 +138,7 @@
     "name": "NgModuleRef"
   },
   {
-    "name": "NodeInjector"
+    "name": "NodeInjector$1"
   },
   {
     "name": "NodeInjectorFactory"
