+++ packages/core/test/bundling/hello_world/bundle.golden_symbols.json
@@ -24,7 +24,7 @@
     "name": "DECLARATION_VIEW"
   },
   {
-    "name": "EMPTY$1"
+    "name": "EMPTY"
   },
   {
     "name": "EMPTY_ARRAY$1"
@@ -48,7 +48,7 @@
     "name": "HOST_NODE"
   },
   {
-    "name": "INJECTOR$1"
+    "name": "INJECTOR"
   },
   {
     "name": "INJECTOR_SIZE"
