+++ packages/core/test/bundling/injection/bundle.golden_symbols.json
@@ -12,7 +12,7 @@
     "name": "EmptyErrorImpl"
   },
   {
-    "name": "INJECTOR"
+    "name": "INJECTOR$1"
   },
   {
     "name": "Inject"
