+++ packages/core/test/render3/ivy/jit_spec.ts
@@ -10,7 +10,7 @@ import 'reflect-metadata';
 
 import {InjectorDef, defineInjectable} from '@angular/core/src/di/defs';
 import {Injectable} from '@angular/core/src/di/injectable';
-import {inject, setCurrentInjector} from '@angular/core/src/di/injector';
+import {inject, setCurrentInjector} from '@angular/core/src/di/injector_compatibility';
 import {ivyEnabled} from '@angular/core/src/ivy_switch';
 import {Component, HostBinding, HostListener, Input, Output, Pipe} from '@angular/core/src/metadata/directives';
 import {NgModule, NgModuleDef} from '@angular/core/src/metadata/ng_module';
