+++ packages/core/test/view/ng_module_spec.ts
@@ -8,7 +8,8 @@
 
 import {NgModuleRef} from '@angular/core';
 import {InjectableDef, defineInjectable} from '@angular/core/src/di/defs';
-import {INJECTOR, InjectFlags, Injector, inject} from '@angular/core/src/di/injector';
+import {INJECTOR, Injector} from '@angular/core/src/di/injector';
+import {InjectFlags, inject} from '@angular/core/src/di/injector_compatibility';
 import {makePropDecorator} from '@angular/core/src/util/decorators';
 import {NgModuleDefinition, NgModuleProviderDef, NodeFlags} from '@angular/core/src/view';
 import {moduleDef, moduleProvideDef, resolveNgModuleDep} from '@angular/core/src/view/ng_module';
